Renaming authorization file to authz details
