+++ cloud-wallet-openid4vc/src/issuance/mod.rs
@@ -1,5 +1,5 @@
 //! OpenID4VCI protocol data models.
-pub mod authz_detail;
+pub mod authz_details;
 pub mod authz_request;
 pub mod authz_response;
 pub mod authz_server_metadata;
